Moved all one of product templates from routes -> templates, for consistency
