+++ src/components/templates/index.server.ts
@@ -1,3 +1,4 @@
+export {ProductTemplate} from './product-template.server';
 export {ProductMetafieldPomadeTemplate} from './product-metafield-pomade-template.server';
 export {ProductMetafieldColorsGetInspiredTemplate} from './product-metafield-colors-get-inspired-template.server';
 export {ProductMetafieldGetInspiredTemplate} from './product-metafield-get-inspired-template.server';
